cephfs: remove CephFS emulation layer
This emulation layer is built on top of the local file system, and is
the basic way to make unit tests work with upstream Hadoop. Keeping it
in sync with the fast moving changes is slowing things down. We intend
to add this back in if and when we decide to shoot for upstream
submission.

Signed-off-by: Noah Watkins <noahwatkins@gmail.com>