cephfs: add harness to run FS tests on live CephFS
File systems typically use an emulation layer that runs the Hadoop
Common file system unit tests on top of a local file system to eliminate
external dependencies. We'd like to run these tests against a running
Ceph instance.

This patch adds a test harness that runs the unit tests against a live
Ceph mount. The configuration is awkward, and we need to figure out how
to integrate this into a workflow. It also is something we won't want to
send to upstream Hadoop.

Relevant config:

  build.xml: set java.library.path
  LD_LIBRARY_PATH: set to relevant dir
  TestCephFileSystem.java: path to ceph.conf

Signed-off-by: Noah Watkins <noahwatkins@gmail.com>