cephfs: port rename to new CephFS interface
Uses rename from the new Java wrappers. All of the unit rename unit
tests pass. We need to watch out for all of the various errors that can
be thrown by CephFS (any errno value), and handle them. Currently we
only watch for the most common problems, but we should use a catch all
exception.

For example, rename() will return ENOENT or ENOTDIR, but in both cases,
Hadoop would expect FileNotFoundException. We can either convert the
later to the former in CephFS, or expose ENOTDIR and do the translation
at the Hadoop level.

Signed-off-by: Noah Watkins <noahwatkins@gmail.com>