cephfs: use Hadoop conf file to setup tests
Rather than pull in several environment variables, instead we run the
tests using a real Hadoop configuration file. This was initially done to
provide several data pool names for the test, but this is also a good
idea because it implicitly tests the configuration options that we
support and which are in the documentation.

Signed-off-by: Noah Watkins <noahwatkins@gmail.com>